requirements.txt
et_xmlfile==2.0.0
lxml==5.4.0
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
six==1.17.0
typing_extensions==4.13.2
tzdata==2025.2
